# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Transaction validation service with rules and limits
  termsOfService: http://swagger.io/terms/
  title: Tracer API
  version: 0.1.0
servers:
- url: http://localhost:8080/
- url: https://localhost:8080/
paths:
...
  /v1/validations:
    get:
      description: Lists transaction validation records with cursor-based pagination
        and filters.
      operationId: listValidations
      parameters:
      - description: "Max items per page (1-1000, default: 100)"
        in: query
        name: limit
        schema:
          maximum: 1000
          minimum: 1
          type: integer
      - description: Pagination cursor from previous response
        in: query
        name: cursor
        schema:
          type: string
      - description: "Field to sort by (default: createdAt)"
        in: query
        name: sortBy
        schema:
          enum:
          - createdAt
          - processingTimeMs
          type: string
      - description: "Sort direction (default: DESC)"
        in: query
        name: sortOrder
        schema:
          enum:
          - ASC
          - DESC
          type: string
      - description: Filter from this date (RFC3339)
        in: query
        name: startDate
        schema:
          format: date-time
          type: string
      - description: Filter to this date (RFC3339)
        in: query
        name: endDate
        schema:
          format: date-time
          type: string
      - description: Filter by decision
        in: query
        name: decision
        schema:
          enum:
          - ALLOW
          - DENY
          - REVIEW
          type: string
      - description: Filter by account ID (UUID)
        in: query
        name: accountId
        schema:
          format: uuid
          type: string
      - description: Filter by matched rule ID (UUID)
        in: query
        name: matchedRuleId
        schema:
          format: uuid
          type: string
      - description: Filter by exceeded limit ID (UUID)
        in: query
        name: exceededLimitId
        schema:
          format: uuid
          type: string
      - description: Filter by segment ID (UUID)
        in: query
        name: segmentId
        schema:
          format: uuid
          type: string
      - description: Filter by portfolio ID (UUID)
        in: query
        name: portfolioId
        schema:
          format: uuid
          type: string
      - description: Filter by transaction type
        in: query
        name: transactionType
        schema:
          enum:
          - CARD
          - WIRE
          - PIX
          - CRYPTO
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http_in.ListTransactionValidationsResponse'
          description: Transaction validations listed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid parameters
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: List transaction validation records
      tags:
      - validations
    post:
      description: Validates a transaction against configured rules and limits.
      operationId: validateTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tracer_pkg_model.ValidationRequest'
        description: Validation request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.ValidationResponse'
          description: Validation result
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Payload too large (exceeds 100KB)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Service unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Gateway timeout
      security:
      - ApiKeyAuth: []
      summary: Validate a transaction
      tags:
      - validations
      x-codegen-request-body-name: request
...
components:
  schemas:
    api.ErrorResponse:
      example:
        code: TRC-0001
        message: Invalid input provided
        title: Bad Request
      properties:
        code:
          example: TRC-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    internal_adapters_http_in.ListTransactionValidationsResponse:
      example:
        nextCursor: nextCursor
        hasMore: true
        transactionValidations:
        - reason: reason
          amount: 0
          validationId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          decision: ALLOW
          matchedRuleIds:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          transactionType: CARD
          accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          createdAt: 2000-01-23T04:56:07.000+00:00
          exceededLimitIds:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          processingTimeMs: 6
          segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          currency: currency
        - reason: reason
          amount: 0
          validationId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          decision: ALLOW
          matchedRuleIds:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          transactionType: CARD
          accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          createdAt: 2000-01-23T04:56:07.000+00:00
          exceededLimitIds:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          processingTimeMs: 6
          segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          currency: currency
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: string
        transactionValidations:
          items:
            $ref: '#/components/schemas/internal_adapters_http_in.ValidationSummary'
          type: array
      type: object
...
    internal_adapters_http_in.ValidationSummary:
      example:
        reason: reason
        amount: 0
        validationId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        decision: ALLOW
        matchedRuleIds:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        transactionType: CARD
        accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        createdAt: 2000-01-23T04:56:07.000+00:00
        exceededLimitIds:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        processingTimeMs: 6
        segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        currency: currency
      properties:
        accountId:
          format: uuid
          type: string
        amount:
          type: integer
        createdAt:
          format: date-time
          type: string
        currency:
          type: string
        decision:
          $ref: '#/components/schemas/tracer_pkg_model.Decision'
        exceededLimitIds:
          items:
            format: uuid
            type: string
          type: array
        matchedRuleIds:
          items:
            format: uuid
            type: string
          type: array
        portfolioId:
          format: uuid
          type: string
        processingTimeMs:
          type: integer
        reason:
          type: string
        segmentId:
          format: uuid
          type: string
        transactionType:
          $ref: '#/components/schemas/tracer_pkg_model.TransactionType'
        validationId:
          format: uuid
          type: string
      type: object
    tracer_pkg_model.AccountContext:
      example:
        accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        metadata:
          key: "{}"
        type: type
        status: status
      properties:
        accountId:
          format: uuid
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        status:
          type: string
        type:
          type: string
      type: object
...
    tracer_pkg_model.Decision:
      enum:
      - ALLOW
      - DENY
      - REVIEW
      type: string
      x-enum-varnames:
      - DecisionAllow
      - DecisionDeny
      - DecisionReview
...
    tracer_pkg_model.LimitUsageDetail:
      example:
        period: period
        exceeded: true
        attemptedAmount: 0
        scope: scope
        currentUsage: 6
        limitAmount: 1
        limitId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        attemptedAmount:
          description: |-
            AttemptedAmount is the transaction amount being validated.
            Per API Design v1.3.2 section 4.1.1.
          type: integer
        currentUsage:
          description: |-
            CurrentUsage represents the PROJECTED usage after applying the transaction amount,
            not the actual persisted counter value. This is calculated as:
            (counter.CurrentUsage + input.Amount) for DAILY/MONTHLY limits, or 0 for PER_TRANSACTION.
            When Exceeded=true, the counter was NOT incremented, but CurrentUsage still shows
            what the usage would have been if the transaction were allowed.
          type: integer
        exceeded:
          type: boolean
        limitAmount:
          type: integer
        limitId:
          format: uuid
          type: string
        period:
          description: |-
            Period indicates the type of limit (DAILY, MONTHLY, PER_TRANSACTION).
            Named "period" per API Design v1.3.2 section 4.1.1.
          type: string
        scope:
          description: |-
            Scope is a human-readable string representation of the limit's scope
            (e.g., "account:uuid" or "segment:uuid" or "global").
            Per API Design v1.3.2 section 4.1.1.
          type: string
      type: object
    tracer_pkg_model.MerchantContext:
      example:
        country: country
        metadata:
          key: "{}"
        merchantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        category: category
      properties:
        category:
          description: "e.g., ISO 18245 MCC code"
          type: string
        country:
          description: ISO 3166-1 alpha-2 code
          type: string
        merchantId:
          format: uuid
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          type: string
      type: object
    tracer_pkg_model.PortfolioContext:
      example:
        metadata:
          key: "{}"
        portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
      properties:
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          type: string
        portfolioId:
          format: uuid
          type: string
      type: object
...
    tracer_pkg_model.SegmentContext:
      example:
        metadata:
          key: "{}"
        segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
      properties:
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          type: string
        segmentId:
          format: uuid
          type: string
      type: object
    tracer_pkg_model.TransactionType:
      enum:
      - CARD
      - WIRE
      - PIX
      - CRYPTO
      type: string
      x-enum-varnames:
      - TransactionTypeCard
      - TransactionTypeWire
      - TransactionTypePix
      - TransactionTypeCrypto
...
    tracer_pkg_model.ValidationRequest:
      properties:
        account:
          $ref: '#/components/schemas/tracer_pkg_model.AccountContext'
        amount:
          type: integer
        currency:
          type: string
        merchant:
          $ref: '#/components/schemas/tracer_pkg_model.MerchantContext'
        metadata:
          additionalProperties:
            type: object
          type: object
        portfolio:
          $ref: '#/components/schemas/tracer_pkg_model.PortfolioContext'
        requestId:
          format: uuid
          type: string
        segment:
          $ref: '#/components/schemas/tracer_pkg_model.SegmentContext'
        subType:
          type: string
        transactionTimestamp:
          format: date-time
          type: string
        transactionType:
          $ref: '#/components/schemas/tracer_pkg_model.TransactionType'
      required:
      - account
      - amount
      - currency
      - requestId
      - transactionTimestamp
      - transactionType
      type: object
    tracer_pkg_model.ValidationResponse:
      example:
        limitUsageDetails:
        - period: period
          exceeded: true
          attemptedAmount: 0
          scope: scope
          currentUsage: 6
          limitAmount: 1
          limitId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - period: period
          exceeded: true
          attemptedAmount: 0
          scope: scope
          currentUsage: 6
          limitAmount: 1
          limitId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        reason: reason
        validationId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        decision: ALLOW
        processingTimeMs: 5
        requestId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        totalRulesLoaded: 5
        truncated: true
        evaluatedRuleIds:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        matchedRuleIds:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        decision:
          $ref: '#/components/schemas/tracer_pkg_model.Decision'
        evaluatedRuleIds:
          items:
            format: uuid
            type: string
          type: array
        limitUsageDetails:
          items:
            $ref: '#/components/schemas/tracer_pkg_model.LimitUsageDetail'
          type: array
        matchedRuleIds:
          items:
            format: uuid
            type: string
          type: array
        processingTimeMs:
          type: integer
        reason:
          type: string
        requestId:
          format: uuid
          type: string
        totalRulesLoaded:
          type: integer
        truncated:
          type: boolean
        validationId:
          format: uuid
          type: string
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: API Key for authentication
      in: header
      name: X-API-Key
      type: apiKey
...